+++ Sources/SwiftUIKit/Data/StandardQrCodeGenerator.swift
@@ -9,24 +9,11 @@
 #if os(iOS) || os(tvOS) || os(macOS)
 import SwiftUI
 
-public class StandardQrCodeGenerator: QrCodeGenerator {
+@available(*, deprecated, message: "Use StandardScanCodeGenerator instead")
+public class StandardQrCodeGenerator: StandardScanCodeGenerator, QrCodeGenerator {
     
-    public init(scale: CGFloat) {
-        transform = CGAffineTransform(scaleX: scale, y: scale)
-    }
-    
-    private let transform: CGAffineTransform
-
     public func generateQRCode(from string: String) -> Image? {
-        let ciContext = CIContext()
-        let data = string.data(using: String.Encoding.utf8)
-        guard let filter = CIFilter(name: "CIQRCodeGenerator") else { return nil }
-        filter.setValue(data, forKey: "inputMessage")
-        guard
-            let ciImage = filter.outputImage?.transformed(by: transform),
-            let cgImage = ciContext.createCGImage(ciImage, from: ciImage.extent)
-            else { return nil }
-        return Image(cgImage, scale: 2, label: Text("QR Code"))
+        generateCodeView(of: .qr, from: string)
     }
 }
 #endif